+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [1.2.0] - 12/10/2019
+* You can do action (pas a closure / callback) at Scroll Ending Event (onScrollEnded)
+
 ## [1.1.0] - 06/10/2019
 * The list scroll to the initial Selected Item
 * Now, the selected item is auto placed on the center